// src/types/whiteboard.ts
// types.ts (or whiteboardApi.ts)
export type WhiteboardTool =
  | 'pen'
  | 'eraser'
  | 'rectangle'
  | 'text'
  | 'bold'
  | 'italic'
  | 'alignLeft'
  | 'alignCenter'
  | 'alignRight'
  | 'list'
  | 'line'
  | 'arrow'
  | 'circle'
  | 'leftArrow'
  | 'downArrow'
  | 'upArrow'
  | 'triangle'
  | 'diamond'
  | 'polygon'
  | 'straightConnector'
  | 'elbowConnector'
  | 'curvedConnector';